Handle multi-chunk messages, better logging
Implement a dual go-routine approach to splicing the client to the
server so data flows more in a duplex fashion. This fixes issues with
large result sets that come across in many chunks that don't fit in
the buffer.

Also clean up logging and make it more debuggable. Looks closer to the
Python driver output now.