Resolve key Bolt message parsing & splicing issues
In order to properly get details like the database, principal, etc. we
need to know how to parse certain types of Bolt message
payloads. Specifically this means handling unpacking the "tinymap"
that is sent with a Run, Begin, and Success message.

Since we also need to detect and switch backend connections on the
start of a new transaction, the original splicing logic was
fundamentaly flawed and wouldn't support switching between read and
write backend servers. Instead of using the splice routine to deal
with both directions of communication, for each new transaction splice
is used only for piping the server's response to the client. This way
the client state is maintained in the same, original handleClient()
routine.

There might be a way to design something more powerful, but it's
already sloppy enough with the message parsing and stuff.

Still only handles a single backend system (by default will now pick a
writer from the routing table) and won't context switch yet based on
the database, but those are the next steps :-)