can't do chunking properly...
This change focuses primarily on dealing with message
chunking. Reworked it to properly rebuild chunks, but the problem is
we're now rebuilding complete messages from chunks and then sending
those. Probably too large for websocket-based messages as Browser is
acting strangely now.

Another oddity is the use of RUN/RESET by some clients (cypher-shell
and Browser). Probably because they try to do a lot of reliance on the
Session.Run() API? Not sure yet, but it seems to break my proxy.