neo4j desktop hanging on remote connection...
Try dealing with a deadlock induced by neo4j desktop connecting but
sending a reset message right away after authenticating. It's odd
behavior I haven't seen elsewhere and since we don't have a server
connection to reset (since we haven't opened a transaction state)
we're currently just sitting and reading both client & server.

So in the event of a RESET and we're not connected to a backend
server, just send an empty SUCCESS message because why not?