hardcoding for v4.1 was silly and meant compat issues
Add in proper version negotiation with the client by knowing the
current server version. Logic is as basically as follows:

  min(max(clientVersion), serverVersion)

Since the client will say which 4 versions it supports, we need to
find the right version of that list to compare with the server.

Also adds a Version type we now track on the Monitor and logs the
backend version at connect. (Might help debugging in the future.)

Lots of printf logging for handshakes for now until I trust it more,
but the unit tests should have caught a few things. Though I admit the
test conditions are not exhaustive.

This seems to work for now since we map 1 client conn to a private
pool of backend conns. If we get to a shared conn pool, will need to
be more cautious...but let's maybe never do that?