it's important to stay healthy in 2020
Add a simple HTTP healthcheck. Or is it "health check"?

Now more than ever I need to clean up some of this code. Put this
healthcheck stuff into a "health" module just to help get it out of
the way for now. I both love having most logic in proxy.go...and hate
it at the same time. I like just running `go build` with no args.

Anyways: you can now send an "HTTP GET /health" to the proxy and it
will report back a 200 OK...if it's alive. No other "checks" are made
yet, but downline can wire in details on the cluster status
potentially. Basically if the proxy can't reach the backend, it's not
worth sending traffic to it!