chunky peanut butter is crap, but chunked messages are ok
Original implementation of readMessage() for direct bolt connections
de-chunked bolt Messages. This was ok for some things, but when a
WebSocket connection wrote a massive chunked payload, it could result
in massive bolt Messages to the Server. Sure, not too big a problem...

But when the Server would reply with a massive response that was
chunked (looks like they're usually on 16kb boundaries?) the
bolt-proxy was de-chunking and causing VERY large Messages to the
client. For WebSocket clients, this was a tad problematic.

This change preserves chunks. We don't care about parsing them most of
the time...so just relay them as-is.

The logic is different than with WebSocket connections because we're
not dealing with another layer of framing, so we just do some wasteful
2-byte reads for now and have a bonus 2 byte buffer to deal with
getting the next chunks msg length prematurely.