bugs in under-read logic causing malformed messages
Some bad bit wrangling and counting was causing some messages to
appear chunked when they weren't...this fixes that and addresses some
cases of under-reads. Still panics if 5 occur in a row. If so,
something is up...and we need a better way to read a net.Conn maybe?

In practice, the under-reads were for 2 byte reads so not sure why the
OS wouldn't be able to fulfill that request. Huh.

Tested with Aura and local single-instance. Things work ok.

Now that they're working, this also decreases the buffer sizes from
128kb to 32kb, which should still be "too much" buffer given I think
the server chunks ~16kb? Not sure but I'd presume drivers do too.