fixed routing table update intervals miss changes
I'm not too thrilled with this change...it feels messier...but it's a
start at fetching Routing Tables on demand and caching them for some
period of time. This solves the problem of a new database being
created and it not being accessible immediately.

There are still some issues for newly created databases and they need
to be ironed out, but existing databases should still behave the same
as before.